requirements.txt
python-telegram-bot
python-dotenv
pydantic-settings
pillow
rembg[cpu]